inline prompt selection
